+++ src/main/java/com/greedy/mokkoji/api/admin/service/AdminService.java
@@ -6,8 +6,8 @@ import com.greedy.mokkoji.api.pagination.dto.PageResponse;
 import com.greedy.mokkoji.common.exception.MokkojiException;
 import com.greedy.mokkoji.db.club.entity.Club;
 import com.greedy.mokkoji.db.club.repository.ClubRepository;
-import com.greedy.mokkoji.db.clubMasterApplication.entity.ClubMasterApplication;
-import com.greedy.mokkoji.db.clubMasterApplication.repository.ClubMasterApplicationRepository;
+import com.greedy.mokkoji.db.clubmaster.entity.ClubMasterApplication;
+import com.greedy.mokkoji.db.clubmaster.repository.ClubMasterApplicationRepository;
 import com.greedy.mokkoji.db.user.entity.User;
 import com.greedy.mokkoji.db.user.repository.UserRepository;
 import com.greedy.mokkoji.enums.auth.AuthRole;
