+++ src/main/java/com/greedy/mokkoji/api/clubmaster/controller/ClubMasterController.java
@@ -1,10 +1,10 @@
-package com.greedy.mokkoji.api.clubMaster.controller;
+package com.greedy.mokkoji.api.clubmaster.controller;
 
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.AuthCredential;
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.Authentication;
-import com.greedy.mokkoji.api.clubMaster.dto.request.CreateClubMasterApplicationsRequest;
-import com.greedy.mokkoji.api.clubMaster.dto.response.GetMyClubMasterApplicationsResponse;
-import com.greedy.mokkoji.api.clubMaster.service.ClubMasterService;
+import com.greedy.mokkoji.api.clubmaster.dto.request.CreateClubMasterApplicationsRequest;
+import com.greedy.mokkoji.api.clubmaster.dto.response.GetMyClubMasterApplicationsResponse;
+import com.greedy.mokkoji.api.clubmaster.service.ClubMasterService;
 import com.greedy.mokkoji.common.response.APISuccessResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
