+++ src/main/java/com/greedy/mokkoji/api/clubmaster/controller/ClubMasterControllerSwagger.java
@@ -1,8 +1,8 @@
-package com.greedy.mokkoji.api.clubMaster.controller;
+package com.greedy.mokkoji.api.clubmaster.controller;
 
 import com.greedy.mokkoji.api.auth.controller.argumentResolver.AuthCredential;
-import com.greedy.mokkoji.api.clubMaster.dto.request.CreateClubMasterApplicationsRequest;
-import com.greedy.mokkoji.api.clubMaster.dto.response.GetMyClubMasterApplicationsResponse;
+import com.greedy.mokkoji.api.clubmaster.dto.request.CreateClubMasterApplicationsRequest;
+import com.greedy.mokkoji.api.clubmaster.dto.response.GetMyClubMasterApplicationsResponse;
 import com.greedy.mokkoji.common.response.APISuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
